refactor(plugin-support): move HelpMenu from plugin-status-bar
The right-rail help dropdown (?) is a user-assistance affordance, not a
status-bar concern. plugin-support already owns the welcome tour,
shortcuts dialog, and feedback panel, so it's the natural home for the
help menu too.

- Move HelpMenu container into plugin-support.
- Contribute the status-indicator surface from plugin-support's
  react-surface module instead of plugin-status-bar's.
- Move help-menu / docs / contact-us / shortcuts / discord / github /
  download-apps / about translation keys to plugin-support's namespace;
  keep beta-callout keys (warning / technology-preview / learn-more /
  released / see-release / powered-by-dxos) in plugin-status-bar where
  the VersionNumber popover still uses them.
- Drop the inlined SHORTCUTS_DIALOG constant: now that the menu lives
  inside plugin-support there's no dep cycle, so import the canonical
  constant from ../../constants. The welcome ABOUT_DIALOG string stays
  inlined because composer-app is not a workspace package.
- Add @dxos/plugin-status-bar dep + tsconfig ref to plugin-support so
  the HelpMenu trigger can still use StatusBar.Item for visual parity
  with the other status indicators.
- Update docs link to https://docs.dxos.org/composer/introduction/.
