+++ packages/plugins/plugin-status-bar/src/translations.ts
@@ -17,15 +17,7 @@ export const translations = [
         'released.message': 'This version was released {{released}}.',
         'see-release.label': 'See release on GitHub',
         'powered-by-dxos.message': 'Powered by <dxos>DXOS</dxos>',
-        'discord.label': 'Discord',
-        'github.label': 'GitHub',
         'feedback.label': 'Feedback',
-        'help-menu.label': 'Help & resources',
-        'docs.label': 'Documentation',
-        'contact-us.label': 'Contact us',
-        'shortcuts.label': 'Keyboard shortcuts',
-        'download-apps.label': 'Download apps',
-        'about.label': 'About Composer',
       },
     },
   },
