+++ packages/plugins/plugin-support/src/containers/index.ts
@@ -5,6 +5,7 @@
 import { type ComponentType, lazy } from 'react';
 
 export const FeedbackPanel: ComponentType<any> = lazy(() => import('./FeedbackPanel'));
+export const HelpMenu: ComponentType<any> = lazy(() => import('./HelpMenu'));
 export const ShortcutsDialogContent: ComponentType<any> = lazy(() => import('./ShortcutsDialogContent'));
 export const ShortcutsHints: ComponentType<any> = lazy(() => import('./ShortcutsHints'));
 export const ShortcutsList: ComponentType<any> = lazy(() => import('./ShortcutsList'));
