+++ packages/plugins/plugin-support/tsconfig.json
@@ -108,6 +108,9 @@
       "path": "../plugin-space"
     },
     {
+      "path": "../plugin-status-bar"
+    },
+    {
       "path": "../plugin-testing"
     },
     {
